fix ton-depool-validation-request.sh
-delete unworked check
-

+++ roles/monitoring_agent/files/scripts/ton-depool-validation-request.sh
@@ -69,12 +69,12 @@ echo "INFO: eclection is active"
 ##=================
 
 #already completed request for election
-PARTICIPANT_STATE=$(ton-node-participant-state.sh)
-if [ $PARTICIPANT_STATE == "ACTIVE" ]
-   then
-        echo "INFO: already in participants list"
-        exit 0
-fi;
+#PARTICIPANT_STATE=$(ton-node-participant-state.sh)
+#if [ $PARTICIPANT_STATE == "ACTIVE" ]
+#   then
+#        echo "INFO: already in participants list"
+#        exit 0
+#fi;
 
 
 ##=================
@@ -123,21 +123,25 @@ TON_PROXY=$(cat $TON_ELECTION_SUBFOLDER/$TON_ELECTION_PROXY_FILE_NAME)
 if [ -z $TON_PROXY ]; then
    echo "ERROR: proxy data is empty $TON_ELECTION_SUBFOLDER/$TON_ELECTION_PROXY_FILE_NAME"
 fi
-TON_CONSOLE_CONFIG_NEW=$(jq ".wallet_id=$TON_PROXY" $TON_CONSOLE_CONFIG)  
-echo $TON_CONSOLE_CONFIG_NEW > $TON_CONSOLE_CONFIG
+
+if [ ! -f $TON_ELECTION_SUBFOLDER/validator-query.boc ]; then
+{
+
+   TON_CONSOLE_CONFIG_NEW=$(jq ".wallet_id=$TON_PROXY" $TON_CONSOLE_CONFIG)  
+   echo $TON_CONSOLE_CONFIG_NEW > $TON_CONSOLE_CONFIG
 
 
-ELECTOR_CONFIG=`$TON_CLI --url $TON_DAPP getconfig 15` 
-ELECTOR_CONFIG_JSON=$(echo $ELECTOR_CONFIG | awk '{split($0, a, "p15:"); print a[2]}')
-ELECTOR_CONFIG_VALIDATORS_ELECTED_FOR=`echo "$ELECTOR_CONFIG_JSON" | jq ".validators_elected_for"`
-ELECTOR_CONFIG_STAKE_HELD_FOR=`echo "$ELECTOR_CONFIG_JSON" | jq ".stake_held_for"`
+   ELECTOR_CONFIG=`$TON_CLI --url $TON_DAPP getconfig 15` 
+   ELECTOR_CONFIG_JSON=$(echo $ELECTOR_CONFIG | awk '{split($0, a, "p15:"); print a[2]}')
+   ELECTOR_CONFIG_VALIDATORS_ELECTED_FOR=`echo "$ELECTOR_CONFIG_JSON" | jq ".validators_elected_for"`
+   ELECTOR_CONFIG_STAKE_HELD_FOR=`echo "$ELECTOR_CONFIG_JSON" | jq ".stake_held_for"`
 
-VALIDATION_START=$ELECTIONS_DATE
-VALIDATION_END=$(($VALIDATION_START + $ELECTOR_CONFIG_VALIDATORS_ELECTED_FOR + $ELECTOR_CONFIG_STAKE_HELD_FOR + 600))
+   VALIDATION_START=$ELECTIONS_DATE
+   VALIDATION_END=$(($VALIDATION_START + $ELECTOR_CONFIG_VALIDATORS_ELECTED_FOR + $ELECTOR_CONFIG_STAKE_HELD_FOR + 600))
 
-$TON_CONSOLE -C $TON_CONSOLE_CONFIG -c "election-bid $VALIDATION_START $VALIDATION_END"
-mv validator-query.boc "${TON_ELECTION_SUBFOLDER}"
-
+   $TON_CONSOLE -C $TON_CONSOLE_CONFIG -c "election-bid $VALIDATION_START $VALIDATION_END"
+   mv validator-query.boc "${TON_ELECTION_SUBFOLDER}"
+fi
 ##=================
 ## region: SEND VALIDATION REQUEST
 ##=================
